to compiling on new FixedLib stubs

+++ contracts/plugins/assets/CompoundPricedAsset.sol
@@ -9,7 +9,7 @@ contract CompoundPricedAsset is CompoundOracleMixin, Asset {
     // solhint-disable no-empty-blocks
     constructor(
         IERC20Metadata erc20_,
-        int192 maxTradeVolume_,
+        uint256 maxTradeVolume_,
         IComptroller comptroller_
     ) CompoundOracleMixin(comptroller_) Asset(erc20_, maxTradeVolume_) {}
